backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Security
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
slowapi==0.1.9

# Settings
pydantic-settings==2.1.0

# Image processing
Pillow==10.2.0

# Excel
openpyxl==3.1.2

# Utils
aiofiles==23.2.1
